refactor(data): Remove WallpaperRepositoryProvider
This commit removes the `WallpaperRepositoryProvider` object.

The responsibility of providing a concrete `WallpaperRepository` implementation (like `WallhavenImageRepository` or `PexelsWallpaperRepository`) based on a selected source is being moved to another layer of the application, simplifying the `WallpaperRepository.kt` file. This change streamlines the repository creation process and decouples it from this specific file.