refactor(UI): Conditionally apply nested scroll behavior
This commit refactors `HomeLayout` to conditionally apply the `nestedScroll` modifier to the `Scaffold`.

The nested scroll behavior, which is responsible for coordinating scrolling effects with the top app bar, is now only enabled for screens that contain scrollable lists (`WallpaperListRoute` and `SearchWallpapersRoute`). This prevents unnecessary scroll listeners from being attached on other screens, such as the settings or about pages.